+++ md-editor/md-editor.ts
@@ -1,0 +1,6 @@
+export interface MdEditorOption {  
+  showBorder: boolean       // Show editor component's border
+  hideIcons: Array<string>  // ['Bold', 'Italic', 'Heading', 'Refrence', 'Link', 'Image', 'Ul', 'Ol', 'Code', 'TogglePreview', 'FullScreen'], Default is empty
+  scrollPastEnd: number     // The option for ace editor
+  enablePreviewContentClick: boolean  // Allow user fire the click event on the preview panel, like href etc.
+}